+++ src/main/java/io/witcradg/ordertrackingapi/persistence/ISalesPersistenceService.java
@@ -1,6 +1,6 @@
 package io.witcradg.ordertrackingapi.persistence;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public interface ISalesPersistenceService {
 	public abstract void write(
@@ -8,6 +8,6 @@ public interface ISalesPersistenceService {
 			String sku,
 			String productName, 
 			Integer quantitySold, 
-			BigInteger unitPrice, 
-			BigInteger totalPrice);
+			BigDecimal unitPrice, 
+			BigDecimal totalPrice);
 }
